+++ templates/pilet-template-ng-standalone/src/versions.ts
@@ -1,14 +1,20 @@
 // Mapping of Angular to TypeScript versions
 export const tsVersions = {
   17: '~5.2',
+  18: '~5.5',
+  19: '~5.7',
 };
 
 // Mapping of Angular to RxJs versions
 export const rxjsVersions = {
   17: '^7.4',
+  18: '^7.4',
+  19: '^7.4',
 };
 
 // Mapping of Angular to Zone.js versions
 export const zoneVersions = {
   17: '0.14.0',
+  18: '~0.14',
+  19: '~0.15',
 };